requirements.txt
prefect==2.20.11
pydantic==2.10.6
pydantic-settings==2.7.1
httpx==0.28.1
psycopg[binary]==3.2.6
sqlalchemy==2.0.35
alembic==1.14.1
minio==7.2.15
mlflow==2.21.0
# Pin CPU-compatible numeric stack for older x86_64 (non-x86-64-v2) hosts.
numpy==1.26.4
pyarrow==14.0.2
orjson==3.10.15
jsonschema==4.23.0
tenacity==9.0.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
